Fixed compile errors and moved some classes to better packages
