+++ src/main/java/use_case/search/SearchInteractor.java
@@ -2,6 +2,7 @@ package use_case.search;
 
 import entity.movie_fields.MovieFieldInterface;
 import entity.movie_fields.MovieFieldRegisterInterface;
+import use_case.MovieDataAccessInterface;
 
 import java.util.Map;
